Team1 bank to Joker's branch

+++ TMS.Net07.Lesson1/SpaceWeb/Controllers/AccountController.cs
@@ -14,7 +14,9 @@ using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using SpaceWeb.Presentation;
 
 namespace SpaceWeb.Controllers
 {
@@ -23,82 +25,71 @@ namespace SpaceWeb.Controllers
     {
         private IBankAccountRepository _bankAccountRepository;
         private IMapper _mapper;
-        private IUserRepository _userRepository; //удалить?
         private IWebHostEnvironment _hostEnvironment;
         private UserService _userService;
+        private IAccountPresentation _accountPresentation;
 
-        public AccountController(IBankAccountRepository bankAccountRepository,
-            QuestionaryRepository profileRepository,
-            IUserRepository userRepository,
-            IMapper mapper, UserService userService, 
-            IWebHostEnvironment hostEnvironment)
+        public AccountController(IBankAccountRepository bankAccountRepository, 
+            IMapper mapper, 
+            IWebHostEnvironment hostEnvironment, 
+            UserService userService, 
+            IAccountPresentation accountPresentation)
         {
             _bankAccountRepository = bankAccountRepository;
             _mapper = mapper;
-            _userService = userService;
             _hostEnvironment = hostEnvironment;
+            _userService = userService;
+            _accountPresentation = accountPresentation;
         }
 
         [HttpGet]
         public IActionResult Index(long id)
         {
-            if (id > 0)
+            if (id <= 0)
             {
-                var user = _userService.GetCurrent();
-                var dbModel = user.BankAccounts?.SingleOrDefault(x => x.Id == id);
-
-                if(dbModel == null)
-                {
-                    return View();
-                }
-
-                var viewModel = _mapper.Map<BankAccountViewModel>(dbModel);
-
-                var index = user.BankAccounts.IndexOf(dbModel);
-
-                viewModel.AccountIndex = index;
-                //для того, чтобы вставить индекс активного аккаунта при загрузке страницы
-
-                return View(viewModel);
+                return RedirectToAction("Creation");
             }
-            return RedirectToAction("Creation");
-        }
 
+            var user = _userService.GetCurrent();
+
+            if (!user.BankAccounts.Any(x => x.Id == id))
+            {
+                return RedirectToAction("Creation");
+            }
+
+            var viewModel = _accountPresentation.GetViewModelForIndex(id);
+
+            return View(viewModel);
+        }
 
         [HttpGet]
-        public IActionResult Remove(long id)
-        {
-            _bankAccountRepository.Remove(id);
-
-            var user = _userService.GetCurrent();
-
-            var newId = user.BankAccounts?.FirstOrDefault()?.Id;
-            if (newId != null)
-            {
-                return Redirect($"/Account/Index?id={newId}");
-            }
-            return RedirectToAction("Creation");
-        }
-
-        [HttpPost]
         public IActionResult Remove(long id, string password)
         {
-            _bankAccountRepository.Remove(id);
-
             var user = _userService.GetCurrent();
 
-            var newId = user.BankAccounts?.FirstOrDefault()?.Id;
-            if (newId != null)
+            if (user.Password != password)
             {
-                return Redirect($"/Account/Index?id={newId}");
+                return Json(false);
             }
-            return RedirectToAction("Creation");
+            else
+            {
+                _bankAccountRepository.Remove(id);
+                var newUrl = ("/Account/Creation");
+                var newId = user.BankAccounts?.FirstOrDefault()?.Id;
+                if (newId != null)
+                {
+                    newUrl = $"/Account/Index?id={newId}";
+                    return Json(newUrl);
+                }
+                return Json(newUrl);
+            }
         }
 
         [HttpGet]
         public IActionResult Creation()
         {
-            return View();
+            var viewModel = _accountPresentation.GetAllViewModelsForCreation();
+            return View(viewModel);
         }
 
         [HttpPost]
@@ -154,7 +145,7 @@ namespace SpaceWeb.Controllers
 
             //return RedirectToAction("Index", new { id });
 
-            return RedirectToRoute("default", new { controller = "Account", action = "Index", id});
+            return RedirectToRoute("default", new { controller = "Account", action = "Index", id });
         }
 
         public IActionResult DownloadAccountsInfo()
@@ -170,9 +161,9 @@ namespace SpaceWeb.Controllers
             var colorNow = 0;
             Border slimLine = new Border(BorderStyle.Tcbs_single, BorderSize.one, 0, Color.Black);
             Border boldLine = new Border(BorderStyle.Tcbs_single, BorderSize.seven, 0, Color.Black);
-            List<Color> colorList = new List<Color>() { 
-                Color.OrangeRed, 
-                Color.CadetBlue, 
+            List<Color> colorList = new List<Color>() {
+                Color.OrangeRed,
+                Color.CadetBlue,
                 Color.LightGreen,
                 Color.PeachPuff,
                 Color.Aqua,
@@ -212,7 +203,7 @@ namespace SpaceWeb.Controllers
                     table.Rows[4].Cells[0].Paragraphs.First().Append("Creation date").Bold().FontSize(14).Italic();
                     table.Rows[5].Cells[0].Paragraphs.First().Append("Expiry date").Bold().FontSize(14).Italic();
 
-                    
+
                     for (int i = 0; i < countRows; i++)
                     {
                         table.Rows[i].Cells[0].FillColor = colorList[colorNow];
@@ -307,9 +298,19 @@ namespace SpaceWeb.Controllers
 
         public IActionResult UpdateAmount(long id, decimal amount)
         {
+
+            var myReg = new Regex(@"[\d]*[.,][\d]{1,2}|[\d]*"); //излишне?
+
+            var isMatch = myReg.IsMatch(amount.ToString());
+
+            if (!isMatch)
+            {
+                return Json(false);
+            }
+
             var account = _bankAccountRepository?.Get(id);
 
-            if(account != null)
+            if (account != null)
             {
                 account.Amount += amount;
                 _bankAccountRepository.Save(account);
@@ -317,6 +318,13 @@ namespace SpaceWeb.Controllers
             }
 
             return Json(false);
+        }
+
+        public IActionResult GetByName(string name)
+        {
+            var userId = _userService.GetCurrent().Id;
+            var answer = _bankAccountRepository.GetByName(userId, name).Select(x => x.Id);
+            return Json(answer);
         }
     }
 }
